WareHouse/WordCloud_THUNLP_20231015215721/requirements.txt
wordcloud
matplotlib
Pillow
